Move files to App
