+++ app/containers/App/index.tsx
@@ -1,3 +1,4 @@
+
 /**
  *
  * App.js
@@ -9,13 +10,25 @@
 
 import * as React from 'react';
 import { Switch, Route } from 'react-router-dom';
+import { compose } from 'redux';
+
+import injectSaga from 'utils/injectSaga';
+import injectReducer from 'utils/injectReducer';
 
 import HomePage from 'containers/HomePage';
 import Editor from 'containers/Editor';
 import Player from 'containers/Player';
 import Viewer from 'containers/Viewer';
+
 import NotFoundPage from 'containers/NotFoundPage/Loadable';
-import { slicerContainer } from 'containers/Slicer';
+
+import reducer from './reducer';
+import saga from './saga';
+
+const withReducer = injectReducer({ key: 'slicer', reducer: reducer });
+const withSaga = injectSaga({ key: 'slicer', saga: saga });
+
+const slicerContainer = compose(withReducer, withSaga);
 
 export default slicerContainer(function App() {
   return (
