--- app/containers/Slicer/Loadable.ts
@@ -1,9 +1,0 @@
-/**
- *
- * Asynchronously loads the component for Slicer
- *
- */
-
-import loadable from 'loadable-components';
-
-export default loadable(() => import('./index'));
