--- app/containers/Slicer/constants.ts
@@ -1,17 +1,0 @@
-/*
- *
- * Slicer constants
- *
- */
-
-enum ActionTypes {
-  SET_PROGRESS = 'app/Slicer/SET_PROGRESS',
-  EXPORT_START = 'app/Slicer/EXPORT_START',
-  EXPORT_SUCCESS = 'app/Slicer/EXPORT_SUCCESS',
-  EXPORT_FAILURE = 'app/Slicer/EXPORT_FAILURE',
-  IMPORT_SLIDESHOW = 'app/Slicer/IMPORT_SLIDESHOW',
-  IMPORT_SLIDESHOW_SUCCESS = 'app/Slicer/IMPORT_SLIDESHOW_SUCCESS',
-  IMPORT_SLIDESHOW_FAILURE = 'app/Slicer/IMPORT_SLIDESHOW_FAILURE',
-}
-
-export default ActionTypes;
