--- app/containers/Slicer/messages.ts
@@ -1,16 +1,0 @@
-/*
- * Slicer Messages
- *
- * This contains all the text for the Slicer container.
- */
-
-import { defineMessages } from 'react-intl';
-
-export const scope = 'app.containers.Slicer';
-
-export default defineMessages({
-  header: {
-    id: `${scope}.header`,
-    defaultMessage: 'This is the Slicer container!',
-  },
-});
