--- app/containers/Slicer/tests/index.test.ts
@@ -1,11 +1,0 @@
-// import React from 'react';
-// import { mount } from 'enzyme';
-// import { enzymeFind } from 'styled-components/test-utils';
-
-// import { Slicer } from '../index';
-
-describe('<Slicer />', () => {
-  it('Expect to have unit tests specified', () => {
-    expect(true).toEqual(false);
-  });
-});
