--- app/containers/Slicer/tests/reducer.test.ts
@@ -1,8 +1,0 @@
-// import { fromJS } from 'immutable';
-// import slicerReducer from '../reducer';
-//
-// describe('slicerReducer', () => {
-//   it('returns the initial state', () => {
-//     expect(slicerReducer(undefined, {})).toEqual(fromJS({}));
-//   });
-// });
