--- app/containers/Slicer/tests/saga.test.ts
@@ -1,15 +1,0 @@
-/**
- * Test sagas
- */
-
-/* eslint-disable redux-saga/yield-effects */
-// import { take, call, put, select } from 'redux-saga/effects';
-// import slicerSaga from '../saga';
-
-// const generator = slicerSaga();
-
-describe('slicerSaga Saga', () => {
-  it('Expect to have unit tests specified', () => {
-    expect(true).toEqual(false);
-  });
-});
