--- app/containers/Slicer/tests/selectors.test.ts
@@ -1,8 +1,0 @@
-// import { fromJS } from 'immutable';
-// import { selectSlicerDomain } from '../selectors';
-
-describe('selectSlicerDomain', () => {
-  it('Expect to have unit tests specified', () => {
-    expect(true).toEqual(false);
-  });
-});
